[RLHF] use worker_extension_cls for compatibility with V0 and V1 (#14185)
